Move Report model to common package.
