+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/Report.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.processor.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.opentestsystem.rdw.reporting.common.report.AbstractExamReportRequest;
