+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/ReportStatus.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.processor.model;
+package org.opentestsystem.rdw.reporting.common.model;
 
 /**
  * An enum of available report statuses.
