+++ pdf-report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/service/ReportService.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.reporting.processor.service;
 
+import org.opentestsystem.rdw.reporting.common.model.Report;
 import org.opentestsystem.rdw.reporting.common.report.BatchExamReportRequestHolder;
-import org.opentestsystem.rdw.reporting.processor.model.Report;
 
 import java.io.InputStream;
 import java.util.Collection;
