+++ pdf-report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/ReportController.java
@@ -2,8 +2,8 @@ package org.opentestsystem.rdw.reporting.processor.web;
 
 import com.amazonaws.services.s3.Headers;
 import org.apache.commons.io.IOUtils;
+import org.opentestsystem.rdw.reporting.common.model.Report;
 import org.opentestsystem.rdw.reporting.common.report.BatchExamReportRequestHolder;
-import org.opentestsystem.rdw.reporting.processor.model.Report;
 import org.opentestsystem.rdw.reporting.processor.service.ReportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
