+++ pdf-report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/ReportControllerIT.java
@@ -5,11 +5,11 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
+import org.opentestsystem.rdw.reporting.common.model.Report;
+import org.opentestsystem.rdw.reporting.common.model.ReportStatus;
 import org.opentestsystem.rdw.reporting.common.report.BatchExamReportRequestHolder;
 import org.opentestsystem.rdw.reporting.common.report.SchoolGradeExamReportRequest;
 import org.opentestsystem.rdw.reporting.common.security.PermissionScope;
-import org.opentestsystem.rdw.reporting.processor.model.Report;
-import org.opentestsystem.rdw.reporting.processor.model.ReportStatus;
 import org.opentestsystem.rdw.reporting.processor.service.ReportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
